+++ src/gnuv2_demangle/src/demangler.rs
@@ -60,9 +60,11 @@ fn demangle_impl<'s>(
     } else if let Some((func_name, args)) = sym.c_split2("__F") {
         demangle_free_function(config, func_name, args)
     } else if let Some((method_name, class_and_args)) =
-        sym.c_split2_r_starts_with("__", |c| matches!(c, '1'..='9' | 'C' | 't' | 'H'))
+        sym.c_split2_r_starts_with("__", |c| matches!(c, '1'..='9' | 'C' | 't'))
     {
         demangle_method(config, method_name, class_and_args)
+    } else if let Some((func_name, s)) = sym.c_split2("__H") {
+        demangle_templated_function(config, func_name, s)
     } else if let Some((func_name, s)) = sym.c_split2("__Q") {
         demangle_namespaced_function(config, func_name, s)
     } else if let Some(sym) = sym.strip_prefix("_vt") {
@@ -78,12 +80,15 @@ fn demangle_destructor<'s>(
     config: &DemangleConfig,
     s: &'s str,
 ) -> Result<String, DemangleError<'s>> {
+    let allow_array_fixup = true;
+
     let (r, namespace, typ) = if let Some(s) = s.strip_prefix('t') {
-        let (r, template, typ) = demangle_template(config, s, &ArgVec::new(config, None))?;
+        let (r, template, typ) =
+            demangle_template(config, s, &ArgVec::new(config, None), allow_array_fixup)?;
         (r, Cow::from(template), Cow::from(typ))
     } else if let Some(s) = s.strip_prefix('Q') {
         let (r, namespaces, trailing_namespace) =
-            demangle_namespaces(config, s, &ArgVec::new(config, None))?;
+            demangle_namespaces(config, s, &ArgVec::new(config, None), allow_array_fixup)?;
         (r, Cow::from(namespaces), Cow::from(trailing_namespace))
     } else {
         let Remaining { r, d: class_name } =
@@ -103,6 +108,7 @@ fn demangle_special<'s>(
     s: &'s str,
     full_sym: &'s str,
 ) -> Result<String, DemangleError<'s>> {
+    let allow_array_fixup = true;
     let c = s
         .chars()
         .next()
@@ -119,13 +125,21 @@ fn demangle_special<'s>(
     } else if let Some(remaining) = s.strip_prefix("ti") {
         return demangle_type_info_node(config, remaining);
     } else if let Some(remaining) = s.strip_prefix('t') {
-        let (remaining, template, typ) =
-            demangle_template(config, remaining, &ArgVec::new(config, None))?;
+        let (remaining, template, typ) = demangle_template(
+            config,
+            remaining,
+            &ArgVec::new(config, None),
+            allow_array_fixup,
+        )?;
 
         (remaining, Some(Cow::from(template)), Cow::from(typ), "")
     } else if let Some(q_less) = s.strip_prefix('Q') {
-        let (remaining, namespaces, trailing_namespace) =
-            demangle_namespaces(config, q_less, &ArgVec::new(config, None))?;
+        let (remaining, namespaces, trailing_namespace) = demangle_namespaces(
+            config,
+            q_less,
+            &ArgVec::new(config, None),
+            allow_array_fixup,
+        )?;
 
         (
             remaining,
@@ -165,6 +179,7 @@ fn demangle_special<'s>(
                             cast,
                             &ArgVec::new(config, None),
                             &ArgVec::new(config, None),
+                            allow_array_fixup,
                         )?
                     else {
                         return Err(DemangleError::UnrecognizedSpecialMethod(op));
@@ -180,10 +195,8 @@ fn demangle_special<'s>(
                         // special symbol, so try to decode as a function instead.
                         if let Some((func_name, args)) = full_sym.c_split2("__F") {
                             demangle_free_function(config, func_name, args)
-                        } else if let Some((incomplete_method_name, class_and_args)) = s
-                            .c_split2_r_starts_with("__", |c| {
-                                matches!(c, '1'..='9' | 'C' | 't' | 'H')
-                            })
+                        } else if let Some((incomplete_method_name, class_and_args)) =
+                            s.c_split2_r_starts_with("__", |c| matches!(c, '1'..='9' | 'C' | 't'))
                         {
                             // split `s` instead of `full_sym` to skip over the
                             // first `__`,
@@ -194,6 +207,8 @@ fn demangle_special<'s>(
 
                             let method_name = &full_sym[..incomplete_method_name.len() + 2];
                             demangle_method(config, method_name, class_and_args)
+                        } else if let Some((func_name, s)) = full_sym.c_split2("__H") {
+                            demangle_templated_function(config, func_name, s)
                         } else {
                             Err(DemangleError::UnrecognizedSpecialMethod(op))
                         }
@@ -211,13 +226,17 @@ fn demangle_special<'s>(
             } = demangle_method_qualifier(remaining);
 
             let (remaining, namespaces) = if let Some(q_less) = remaining.strip_prefix('Q') {
-                let (remaining, namespaces, _trailing_namespace) =
-                    demangle_namespaces(config, q_less, &ArgVec::new(config, None))?;
+                let (remaining, namespaces, _trailing_namespace) = demangle_namespaces(
+                    config,
+                    q_less,
+                    &ArgVec::new(config, None),
+                    allow_array_fixup,
+                )?;
 
                 (remaining, Cow::from(namespaces))
             } else if let Some(r) = remaining.strip_prefix('t') {
                 let (remaining, template, _typ) =
-                    demangle_template(config, r, &ArgVec::new(config, None))?;
+                    demangle_template(config, r, &ArgVec::new(config, None), allow_array_fixup)?;
 
                 (remaining, Cow::from(template))
             } else {
@@ -240,6 +259,7 @@ fn demangle_special<'s>(
             remaining,
             class_name.as_deref(),
             &ArgVec::new(config, None),
+            allow_array_fixup,
         )?
     };
 
@@ -256,7 +276,15 @@ fn demangle_free_function<'s>(
     func_name: &'s str,
     args: &'s str,
 ) -> Result<String, DemangleError<'s>> {
-    let argument_list = demangle_argument_list(config, args, None, &ArgVec::new(config, None))?;
+    let allow_array_fixup = true;
+
+    let argument_list = demangle_argument_list(
+        config,
+        args,
+        None,
+        &ArgVec::new(config, None),
+        allow_array_fixup,
+    )?;
 
     Ok(format!("{func_name}({argument_list})"))
 }
@@ -266,72 +294,30 @@ fn demangle_method<'s>(
     method_name: &'s str,
     class_and_args: &'s str,
 ) -> Result<String, DemangleError<'s>> {
+    let allow_array_fixup = true;
     let Remaining {
         r: remaining,
         d: suffix,
     } = demangle_method_qualifier(class_and_args);
 
     let (remaining, namespace) = if let Some(templated) = remaining.strip_prefix('t') {
-        let (remaining, template, _typ) =
-            demangle_template(config, templated, &ArgVec::new(config, None))?;
+        let (remaining, template, _typ) = demangle_template(
+            config,
+            templated,
+            &ArgVec::new(config, None),
+            allow_array_fixup,
+        )?;
 
         (remaining, Cow::from(template))
     } else if let Some(q_less) = remaining.strip_prefix('Q') {
-        let (remaining, namespaces, _trailing_namespace) =
-            demangle_namespaces(config, q_less, &ArgVec::new(config, None))?;
+        let (remaining, namespaces, _trailing_namespace) = demangle_namespaces(
+            config,
+            q_less,
+            &ArgVec::new(config, None),
+            allow_array_fixup,
+        )?;
 
         (remaining, Cow::from(namespaces))
-    } else if let Some(with_return_type) = remaining.strip_prefix('H') {
-        // Templated function.
-
-        let (remaining, template_args, typ) =
-            demangle_template_with_return_type(config, with_return_type)?;
-
-        let (remaining, typ) = if let Some(typ) = typ {
-            (remaining, Some(typ))
-        } else if let Some(r) = remaining.strip_prefix('t') {
-            let (r, template, _typ) = demangle_template(config, r, &ArgVec::new(config, None))?;
-
-            (r, Some(Cow::from(template)))
-        } else {
-            (remaining, None)
-        };
-
-        let (remaining, argument_list) =
-            demangle_argument_list_impl(config, remaining, typ.as_deref(), &template_args, false)?;
-        let (prefix, array_qualifiers) = if let Some(r) = remaining.strip_prefix('_') {
-            let (r, DemangledArg::Plain(mut ret_type, array_qualifiers)) = demangle_argument(
-                config,
-                r,
-                &ArgVec::new(config, typ.as_deref()),
-                &template_args,
-            )?
-            else {
-                return Err(DemangleError::MalformedTemplateWithReturnTypeMissingReturnType(r));
-            };
-            if !r.is_empty() {
-                return Err(
-                    DemangleError::TrailingDataAfterReturnTypeOfMalformedTemplateWithReturnType(r),
-                );
-            }
-            ret_type.push(' ');
-            (ret_type, array_qualifiers)
-        } else {
-            return Err(DemangleError::MalformedTemplateWithReturnTypeMissingReturnType(remaining));
-        };
-
-        let template_args = template_args.join();
-        let formated_template_args = if template_args.ends_with('>') {
-            format!("<{} >", template_args)
-        } else {
-            format!("<{}>", template_args)
-        };
-        let argument_list = argument_list.join();
-        return Ok(if let Some(typ) = typ {
-            format!("{prefix}{array_qualifiers}{typ}::{method_name}{formated_template_args}({argument_list}){suffix}")
-        } else {
-            format!("{prefix}{array_qualifiers}{method_name}{formated_template_args}({argument_list}){suffix}")
-        });
     } else {
         let Remaining { r, d: class_name } =
             demangle_custom_name(remaining, DemangleError::InvalidClassNameOnMethod)?.d_as_cow();
@@ -347,6 +333,7 @@ fn demangle_method<'s>(
             remaining,
             Some(&namespace),
             &ArgVec::new(config, None),
+            allow_array_fixup,
         )?
     };
 
@@ -355,13 +342,104 @@ fn demangle_method<'s>(
     ))
 }
 
+/// Templated functions and methods.
+///
+/// A templated method is templated individually, it doesn't matter if the
+/// class it comes from is templated or not.
+fn demangle_templated_function<'s>(
+    config: &DemangleConfig,
+    func_name: &'s str,
+    s: &'s str,
+) -> Result<String, DemangleError<'s>> {
+    // Arrays do need to be fixed up if it appears in the template list, but
+    // not in the rest of the definition.
+    let allow_array_fixup = true;
+    let (remaining, template_args, typ) =
+        demangle_template_with_return_type(config, s, allow_array_fixup)?;
+    let allow_array_fixup = false;
+
+    let Remaining {
+        r: remaining,
+        d: suffix,
+    } = demangle_method_qualifier(remaining);
+
+    let (remaining, typ) = if let Some(typ) = typ {
+        (remaining, Some(typ))
+    } else if remaining.starts_with(|c| matches!(c, '1'..='9')) {
+        let Remaining { r, d: namespace } = demangle_custom_name(
+            remaining,
+            DemangleError::InvalidNamespaceOnTemplatedFunction,
+        )?
+        .d_as_cow();
+        (r, Some(namespace))
+    } else if let Some(r) = remaining.strip_prefix('t') {
+        let (r, template, _typ) =
+            demangle_template(config, r, &ArgVec::new(config, None), allow_array_fixup)?;
+
+        (r, Some(Cow::from(template)))
+    } else if let Some(r) = remaining.strip_prefix('Q') {
+        let (r, namespaces, _trailing_namespace) =
+            demangle_namespaces(config, r, &ArgVec::new(config, None), allow_array_fixup)?;
+
+        (r, Some(Cow::from(namespaces)))
+    } else {
+        (remaining, None)
+    };
+
+    let (remaining, argument_list) = demangle_argument_list_impl(
+        config,
+        remaining,
+        typ.as_deref(),
+        &template_args,
+        false,
+        allow_array_fixup,
+    )?;
+    let (return_type, array_qualifiers) = if let Some(r) = remaining.strip_prefix('_') {
+        let (r, DemangledArg::Plain(ret_type, array_qualifiers)) = demangle_argument(
+            config,
+            r,
+            &ArgVec::new(config, typ.as_deref()),
+            &template_args,
+            allow_array_fixup,
+        )?
+        else {
+            return Err(DemangleError::MalformedTemplateWithReturnTypeMissingReturnType(r));
+        };
+        if !r.is_empty() {
+            return Err(
+                DemangleError::TrailingDataAfterReturnTypeOfMalformedTemplateWithReturnType(r),
+            );
+        }
+        // ret_type.push(' ');
+        (ret_type, array_qualifiers)
+    } else {
+        return Err(DemangleError::MalformedTemplateWithReturnTypeMissingReturnType(remaining));
+    };
+
+    let template_args = template_args.join();
+    let formated_template_args = if template_args.ends_with('>') {
+        format!("<{} >", template_args)
+    } else {
+        format!("<{}>", template_args)
+    };
+    let argument_list = argument_list.join();
+    let out = if let Some(typ) = typ {
+        format!("{return_type}{array_qualifiers} {typ}::{func_name}{formated_template_args}({argument_list}){suffix}")
+    } else {
+        format!("{return_type}{array_qualifiers} {func_name}{formated_template_args}({argument_list}){suffix}")
+    };
+    Ok(out)
+}
+
 fn demangle_namespaced_function<'s>(
     config: &DemangleConfig,
     func_name: &'s str,
     s: &'s str,
 ) -> Result<String, DemangleError<'s>> {
+    let allow_array_fixup = true;
+
     let (remaining, namespaces, _trailing_namespace) =
-        demangle_namespaces(config, s, &ArgVec::new(config, None))?;
+        demangle_namespaces(config, s, &ArgVec::new(config, None), allow_array_fixup)?;
 
     let argument_list = if remaining.is_empty() {
         "void"
@@ -371,6 +449,7 @@ fn demangle_namespaced_function<'s>(
             remaining,
             Some(&namespaces),
             &ArgVec::new(config, None),
+            allow_array_fixup,
         )?
     };
 
@@ -382,11 +461,14 @@ fn demangle_type_info_function<'s>(
     config: &DemangleConfig,
     s: &'s str,
 ) -> Result<String, DemangleError<'s>> {
+    let allow_array_fixup = true;
+
     if let (remaining, DemangledArg::Plain(demangled_type, array_qualifiers)) = demangle_argument(
         config,
         s,
         &ArgVec::new(config, None),
         &ArgVec::new(config, None),
+        allow_array_fixup,
     )? {
         if remaining.is_empty() {
             Ok(format!(
@@ -404,11 +486,14 @@ fn demangle_type_info_node<'s>(
     config: &DemangleConfig,
     s: &'s str,
 ) -> Result<String, DemangleError<'s>> {
+    let allow_array_fixup = true;
+
     if let (remaining, DemangledArg::Plain(demangled_type, array_qualifiers)) = demangle_argument(
         config,
         s,
         &ArgVec::new(config, None),
         &ArgVec::new(config, None),
+        allow_array_fixup,
     )? {
         if remaining.is_empty() {
             Ok(format!("{demangled_type}{array_qualifiers} type_info node"))
@@ -424,6 +509,7 @@ fn demangle_virtual_table<'s>(
     config: &DemangleConfig,
     s: &'s str,
 ) -> Result<String, DemangleError<'s>> {
+    let allow_array_fixup = true;
     let mut remaining = s;
     let mut stuff = Vec::new();
 
@@ -433,13 +519,14 @@ fn demangle_virtual_table<'s>(
             .ok_or(DemangleError::VTableMissingDollarSeparator(remaining))?;
 
         remaining = if let Some(r) = remaining.strip_prefix('t') {
-            let (r, template, _typ) = demangle_template(config, r, &ArgVec::new(config, None))?;
+            let (r, template, _typ) =
+                demangle_template(config, r, &ArgVec::new(config, None), allow_array_fixup)?;
 
             stuff.push(Cow::from(template));
             r
         } else if let Some(r) = remaining.strip_prefix('Q') {
             let (r, namespaces, _trailing_namespace) =
-                demangle_namespaces(config, r, &ArgVec::new(config, None))?;
+                demangle_namespaces(config, r, &ArgVec::new(config, None), allow_array_fixup)?;
 
             stuff.push(Cow::from(namespaces));
             r
@@ -461,17 +548,20 @@ fn demangle_namespaced_global<'s>(
     s: &'s str,
     name: &'s str,
 ) -> Result<String, DemangleError<'s>> {
+    let allow_array_fixup = true;
+
     let Some(remaining) = s.strip_prefix('_') else {
         return Err(DemangleError::InvalidNamespacedGlobal(s, name));
     };
 
     let (r, space) = if let Some(r) = remaining.strip_prefix('t') {
-        let (r, template, _typ) = demangle_template(config, r, &ArgVec::new(config, None))?;
+        let (r, template, _typ) =
+            demangle_template(config, r, &ArgVec::new(config, None), allow_array_fixup)?;
 
         (r, Cow::from(template))
     } else if let Some(r) = remaining.strip_prefix('Q') {
         let (r, namespaces, _trailing_namespace) =
-            demangle_namespaces(config, r, &ArgVec::new(config, None))?;
+            demangle_namespaces(config, r, &ArgVec::new(config, None), allow_array_fixup)?;
 
         (r, Cow::from(namespaces))
     } else {