+++ CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix emitting empty parenthesis on non-pointer arrays types.
 - Fix incorrect syntax used for arrays as return types in function pointers.
 - Fix failing to demangle templated functions with large integer values.
+- Fix demangling templated functions inside classes and namespaces.
+- Fix array sizes in templated functions.
 
 ## [0.1.0] - 2025-09-18